A form that clicks the mouse.
This is the first Windows Form program I did, it's a simple program I made to click the wifi button on my laptop on and off so that I can use my night data to download without having to stay up until midnight. The program has two values the user enters, these being the times until clicks. Once started the program counts down and then executes a left click once the timer reaches 0. It also shakes the mouse before to turn the screen back on.